Fix .main-trackInfo-artists not targetting the correct element
For context - seemingly after a Spotify update, .main-trackInfo-artists is not used for the currently playing song artist list (noticed since a plugin adding the next song as well still uses this class, and the date was misplaced). Replaced with classes that work, although might change in the future after a Spotify update.